+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJD7896537CCGF40WG6DT1RC"
+        "01KJF6K6SNJ3ZPZN62G24NFZGT"
     )
     r"""The unique identifier of the data source"""