+++ packages/orq-rc/RELEASES.md
@@ -6199,3 +6199,13 @@ Based on:
 - [python v4.5.0-rc.14] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.14] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.14 - packages/orq-rc
+
+## 2026-02-27 09:26:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.731.2 (2.845.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.15] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.15] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.15 - packages/orq-rc